Time: 20 ms (35.81%) | Memory: 47.3 MB (25.64%) - LeetSync

+++ 943-sum-of-subarray-minimums/sum-of-subarray-minimums.cpp
@@ -30,10 +30,10 @@ public:
             stk.push(i);
         }
 
-        long long sum = 0;
+        ll sum = 0;
 
         for (int i = 0; i < length; ++i) {
-            sum += static_cast<ll>(i - left[i]) * (right[i] - i) * nums[i] % MOD;
+            sum += (ll)(i - left[i]) * (right[i] - i) * nums[i] % MOD;
             sum %= MOD;
         }
